move LoginModal to Modal and rename to Login
